requirements.txt
absl-py
flax
jax
numpy
optax
sentencepiece
tensorflow_datasets
tf-nightly